Full review: fix what it found and verify the rest against the sources
Ant: "review the entire mod for bugs and logic errors", and "fix any
issues you find along the way". All 26 production files were read end to
end. Fixed:
- Texture2DReadPatches.Served was an int counting every pixel read. A mod
  calling GetPixel per pixel adds millions per texture, so it could wrap
  negative in a long session. It is now long.
- FailedAudioClipGuard created its silent clip with AudioClip.Create,
  which is main-thread only, and without a guard. If it threw, the
  constructor went on to read the failed clip's length natively, the
  crash the guard exists to stop. It now hands the constructor null
  instead, a managed exception.
- "Sweep now" gave no feedback at the default report level. A button
  press now always answers.
- The diagnostic report and the settings page now lead with the startup
  checks.
- StartupCheck.Snapshot fields became properties; the Mono harness went
  from 19 CS0649 warnings to 0.

Verified correct, with the source as evidence:
- The orphan sweep's source extensions (png/jpg/jpeg) match Image Opt's
  own cleanup list. Image Opt sends .psd straight to vanilla loading, so
  no valid cache can be swept.
- The FGL settings check names all exist in FasterGameLoadingSettings.
- The VEF and Worldbuilder guards already latch readiness, so per frame
  they cost one flag read.
- VehicleReadback's snapshot, ownership hand-off and DestroyImmediate
  scope, and the per-holder latch in the AnyContentLoaded postfix.

159 + 70 tests, 32 + 3 Mono checks, 0 warnings in the mod and the harness.

+++ Source/ImageOptCompat/Texture2DReadPatches.cs
@@ -35,7 +35,9 @@ internal static class Texture2DReadPatches
     // when a weak source key is collected; never take ownership of Image Opt's sources.
     private static readonly HashSet<Texture2D> OwnedCopies = new();
 
-    internal static int Served { get; private set; }
+    /// long, not int: a mod calling GetPixel per pixel adds millions per texture, and an int wrapped
+    /// negative within a long session.
+    internal static long Served { get; private set; }
 
     /// Whether Image Opt's texture record was found, for the startup check. Resolving it early is
     /// harmless: the lookup is the same one the first read would make.
